necessary edits

+++ src/main/java/Engines/SimpleObjects/TextInformations.java
@@ -11,20 +11,19 @@ import java.util.HashMap;
  */
 public class TextInformations 
 {
-	//Store creation and resource info
 	private String resource_name = "";
 	private LocalDateTime generation_date;
 	
 	private String nif_path = "";
 	private double canBeUsed = Double.NaN;
 	private HashMap<String, Double> metrics_GERBIL = new HashMap<String, Double>();				//USED
-	private HashMap<String, Integer> words_distribution = new HashMap<String, Integer>();		//USED
-	private HashMap<Integer, Integer> words_occurr_distr = new HashMap<Integer, Integer>();		//USED
-	private HashMap<Integer, Integer> symbol_sent_dist = new HashMap<Integer, Integer>();		//USED
-	private HashMap<Integer, Integer> annotation_dist = new HashMap<Integer, Integer>();		//USED
-	private HashMap<String, Integer> syn_error_dist = new HashMap<String, Integer>();			//USED
-	private HashMap<String, Integer> pos_tags_dist = new HashMap<String, Integer>();			//USED
-	HashMap<Character, Integer> symbol_error_dist = new HashMap<Character, Integer>();			//USED
+	private HashMap<String, Double> words_distribution = new HashMap<String, Double>();		//USED
+	private HashMap<Integer, Double> words_occurr_distr = new HashMap<Integer, Double>();		//USED
+	private HashMap<Integer, Double> symbol_sent_dist = new HashMap<Integer, Double>();		//USED
+	private HashMap<Integer, Double> annotation_dist = new HashMap<Integer, Double>();		//USED
+	private HashMap<String, Double> syn_error_dist = new HashMap<String, Double>();			//USED
+	private HashMap<String, Double> pos_tags_dist = new HashMap<String, Double>();			//USED
+	HashMap<Character, Double> symbol_error_dist = new HashMap<Character, Double>();			//USED
 	
 	/**
 	 * This constructor store the original resource file name
@@ -72,11 +71,11 @@ public class TextInformations
 		return metrics_GERBIL;
 	}
 	
-	public HashMap<Integer, Integer> getAnnotation_dist() {
+	public HashMap<Integer, Double> getAnnotation_dist() {
 		return annotation_dist;
 	}
 
-	public void setAnnotation_dist(HashMap<Integer, Integer> annotation_dist) {
+	public void setAnnotation_dist(HashMap<Integer, Double> annotation_dist) {
 		this.annotation_dist = annotation_dist;
 	}
 
@@ -84,51 +83,51 @@ public class TextInformations
 		this.metrics_GERBIL = metrics_GERBIL;
 	}
 
-	public HashMap<String, Integer> getWords_distribution() {
+	public HashMap<String, Double> getWords_distribution() {
 		return words_distribution;
 	}
 
-	public void setWords_distribution(HashMap<String, Integer> words_distribution) {
+	public void setWords_distribution(HashMap<String, Double> words_distribution) {
 		this.words_distribution = words_distribution;
 	}
 
-	public HashMap<Integer, Integer> getWords_occurr_distr() {
+	public HashMap<Integer, Double> getWords_occurr_distr() {
 		return words_occurr_distr;
 	}
 
-	public void setWords_occurr_distr(HashMap<Integer, Integer> words_occurr_distr) {
+	public void setWords_occurr_distr(HashMap<Integer, Double> words_occurr_distr) {
 		this.words_occurr_distr = words_occurr_distr;
 	}
 
-	public HashMap<String, Integer> getSyn_error_dist() {
+	public HashMap<String, Double> getSyn_error_dist() {
 		return syn_error_dist;
 	}
 
-	public void setSyn_error_dist(HashMap<String, Integer> syn_error_dist) {
+	public void setSyn_error_dist(HashMap<String, Double> syn_error_dist) {
 		this.syn_error_dist = syn_error_dist;
 	}
 
-	public HashMap<String, Integer> getPos_tags_dist() {
+	public HashMap<String, Double> getPos_tags_dist() {
 		return pos_tags_dist;
 	}
 
-	public void setPos_tags_dist(HashMap<String, Integer> pos_tags_dist) {
+	public void setPos_tags_dist(HashMap<String, Double> pos_tags_dist) {
 		this.pos_tags_dist = pos_tags_dist;
 	}
 
-	public HashMap<Character, Integer> getSymbol_error_dist() {
+	public HashMap<Character, Double> getSymbol_error_dist() {
 		return symbol_error_dist;
 	}
 
-	public void setSymbol_error_dist(HashMap<Character, Integer> symbol_error_dist) {
+	public void setSymbol_error_dist(HashMap<Character, Double> symbol_error_dist) {
 		this.symbol_error_dist = symbol_error_dist;
 	}
 
-	public HashMap<Integer, Integer> getSymbol_sent_dist() {
+	public HashMap<Integer, Double> getSymbol_sent_dist() {
 		return symbol_sent_dist;
 	}
 
-	public void setSymbol_sent_dist(HashMap<Integer, Integer> symbol_sent_dist) {
+	public void setSymbol_sent_dist(HashMap<Integer, Double> symbol_sent_dist) {
 		this.symbol_sent_dist = symbol_sent_dist;
 	}	
 }
